Exclude Release files from checksumming operation
First, it's useless because we don't have a checksum for them anyway,
and it also could happen that the scan was requested while the file was
open somewhere up in the stack, which caused a deadlock in the mmap
truncation protector.

This is a stop-gap solution, the real one will probably stop using mmap.