Explicitly use set -e in scripts
P: apt-cacher-ng: maintainer-script-without-set-e control/config
N:
N:   The maintainer script passes -e to the shell on the #! line rather than using set -e in the body of the script. This is fine for normal operation, but if the script is run by hand with sh /path/to/script (common in debugging), -e will not be in effect. It's
N:   therefore better to use set -e in the body of the script.
N:
N:   Please refer to Debian Policy Manual section 10.4 for details.