Dl job flow fixes and improvements
Stop using the key index, does not make much sense and the costs of
key and range handling probably outweighs the benefits for the small
number of streams.

Also fix potential memory leaks in the specialized queue.

Simplified ping-pong of the dispatched jobs, not creating a complete
sorted list every time and then putting the remainders back, now taking
straight from the backlog and moving the unhandled items to a prio_queue
immediately.